+++ src/test/java/bio/overture/ego/utils/web/BasicWebResource.java
@@ -1,6 +1,5 @@
-package bio.overture.ego.utils.web.second;
+package bio.overture.ego.utils.web;
 
-import bio.overture.ego.utils.web.ResponseOption;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.http.ResponseEntity;
 
@@ -11,7 +10,9 @@ public class BasicWebResource<T> extends AbstractWebResource<T, ResponseOption<T
     super(restTemplate, serverUrl, responseType);
   }
 
-  @Override protected ResponseOption<T> createResponseOption(ResponseEntity<T> responseEntity) {
+  @Override
+  protected ResponseOption<T> createResponseOption(ResponseEntity<T> responseEntity) {
     return new ResponseOption<>(responseEntity);
   }
+
 }
