+++ src/test/java/bio/overture/ego/utils/web/StringResponseOption.java
@@ -15,7 +15,6 @@ import static bio.overture.ego.utils.Collectors.toImmutableList;
 import static bio.overture.ego.utils.Collectors.toImmutableSet;
 import static bio.overture.ego.utils.Streams.stream;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.springframework.http.HttpStatus.OK;
 
 public class StringResponseOption extends ResponseOption<String>{
 
@@ -26,7 +25,7 @@ public class StringResponseOption extends ResponseOption<String>{
   }
 
   public <R> R extractOneEntity(@NonNull Class<R> entityClass){
-    return assertOk().assertHasBody().map(x -> extractOneEntityFromResponse(x, entityClass));
+    return assertOk().assertHasBody().map(x -> internalExtractOneEntityFromResponse(x, entityClass));
   }
 
   public <R> ListAssert<R> assertPageResultsOfType(Class<R> entityClass){
@@ -36,17 +35,17 @@ public class StringResponseOption extends ResponseOption<String>{
   public <R> List<R> extractPageResults(@NonNull Class<R> entityClass){
     return assertOk()
         .assertHasBody()
-        .map(x -> extractPageResultSetFromResponse(x, entityClass));
+        .map(x -> internalExtractPageResultSetFromResponse(x, entityClass));
   }
 
   public <R> Set<R> extractManyEntities(@NonNull Class<R> entityClass){
     return assertOk()
         .assertHasBody()
-        .map(x -> extractManyEntitiesFromResponse(x, entityClass));
+        .map(x -> internalExtractManyEntitiesFromResponse(x, entityClass));
   }
 
   @SneakyThrows
-  public static <T> List<T> internalExtractPageResultSetFromResponse(ResponseEntity<String> r, Class<T> tClass) {
+  private static <T> List<T> internalExtractPageResultSetFromResponse(ResponseEntity<String> r, Class<T> tClass) {
     val page = MAPPER.readTree(r.getBody());
     assertThat(page).isNotNull();
     return stream(page.path("resultSet").iterator())
@@ -55,36 +54,15 @@ public class StringResponseOption extends ResponseOption<String>{
   }
 
   @SneakyThrows
-  public static <T> T internalExtractOneEntityFromResponse(ResponseEntity<String> r, Class<T> tClass) {
+  private static <T> T internalExtractOneEntityFromResponse(ResponseEntity<String> r, Class<T> tClass) {
     return MAPPER.readValue(r.getBody(), tClass);
   }
 
   @SneakyThrows
-  public static <T> Set<T> internalExtractManyEntitiesFromResponse(ResponseEntity<String> r, Class<T> tClass) {
+  private static <T> Set<T> internalExtractManyEntitiesFromResponse(ResponseEntity<String> r, Class<T> tClass) {
     return Streams.stream(MAPPER.readTree(r.getBody()).iterator())
         .map(x -> MAPPER.convertValue(x, tClass))
         .collect(toImmutableSet());
-  }
-
-  @SneakyThrows
-  public static <T> List<T> extractPageResultSetFromResponse(ResponseEntity<String> r, Class<T> tClass) {
-    assertThat(r.getStatusCode()).isEqualTo(OK);
-    assertThat(r.getBody()).isNotNull();
-    return internalExtractPageResultSetFromResponse(r, tClass);
-  }
-
-  @SneakyThrows
-  public static <T> T extractOneEntityFromResponse(ResponseEntity<String> r, Class<T> tClass) {
-    assertThat(r.getStatusCode()).isEqualTo(OK);
-    assertThat(r.getBody()).isNotNull();
-    return internalExtractOneEntityFromResponse(r, tClass);
-  }
-
-  @SneakyThrows
-  public static <T> Set<T> extractManyEntitiesFromResponse(ResponseEntity<String> r, Class<T> tClass) {
-    assertThat(r.getStatusCode()).isEqualTo(OK);
-    assertThat(r.getBody()).isNotNull();
-    return internalExtractManyEntitiesFromResponse(r, tClass);
   }
 
 }
